+++ tests/FIXTURES/exclude_multischema/expected.sql
@@ -1,0 +1,7 @@
+create table "schema2"."y" (
+    "id" uuid,
+    "value" text
+);
+
+
+alter table "schema1"."t" add column "name" text;